+++ src/player.h
@@ -5,6 +5,8 @@
 #include "quadtree.h"
 #include "bullet.h"
 
+struct AppState;
+
 struct Player {
     int bullets_fired;
     char wasd;
@@ -12,12 +14,13 @@ struct Player {
     SDL_FRect rect;
 };
 
-void initialize_player(struct Player *p, SDL_Renderer *renderer);
+void initialize_player(struct Player *p, struct AppState *as, SDL_Renderer *renderer);
 void handle_input(SDL_Event *e, struct Player *p);
 void update_player_movement(struct Player *p);
 unsigned int check_player_bullet_collision(struct Player *player,
                                            struct Bullet **bullets,
                                            struct QTNode *q_tree,
                                            SDL_Renderer *renderer);
+Uint32 fire_weapon(void *as, SDL_TimerID id, Uint32 interval);
 
 #endif
